add pytest fixture for label times
